django/rest/requirements.txt
coreapi==2.0.9
Django==1.10.2
djangorestframework==3.5.2
httpie==0.9.6
itypes==1.1.0
postgres==2.2.1
psycopg2==2.6.2
Pygments==2.1.3
PyYAML==3.12
requests==2.11.1
uritemplate==3.0.0
